ANDKR | more 아이콘 sag 파일 수정, Analyze -> inspect code

+++ dailyHOTEL/app/src/main/java/com/twoheart/dailyhotel/screen/gourmet/payment/GourmetPaymentLayout.java
@@ -760,7 +760,7 @@ public class GourmetPaymentLayout extends BaseLayout implements View.OnClickList
             @Override
             public void onAnimationUpdate(ValueAnimator animation)
             {
-                int value = ((Integer) animation.getAnimatedValue()).intValue();
+                int value = (Integer) animation.getAnimatedValue();
 
                 mAnimationValue = value;
